moving getRiskAnalys() from common to dataPreparationService
