+++ utils/commons.ts
@@ -5,88 +5,7 @@ import { AxiosResponse } from "axios";
 import { CreatedResource } from "../api/models";
 import { TenantType, Role, SessionTokens } from "../features/common-steps";
 import { apiClient } from "../api";
-import { dataPreparationService } from "../services/data-preparation.service";
 import { generateSessionTokens } from "./session-tokens";
-
-type RiskAnalysisTemplateType = "PA" | "Privato/GSP";
-
-const RISK_ANALYSIS_DATA: Record<
-  RiskAnalysisTemplateType,
-  { completed: unknown; uncompleted: unknown }
-> = {
-  "Privato/GSP": {
-    completed: {
-      purpose: ["INSTITUTIONAL"],
-      institutionalPurpose: ["test"],
-      usesPersonalData: ["NO"],
-      usesThirdPartyPersonalData: ["NO"],
-    },
-    uncompleted: {
-      purpose: ["INSTITUTIONAL"],
-      usesPersonalData: ["NO"],
-      usesThirdPartyPersonalData: ["NO"],
-    },
-  },
-  PA: {
-    completed: {
-      purpose: ["INSTITUTIONAL"],
-      institutionalPurpose: ["test"],
-      personalDataTypes: ["WITH_NON_IDENTIFYING_DATA"],
-      legalBasis: ["CONSENT"],
-      knowsDataQuantity: ["NO"],
-      deliveryMethod: ["CLEARTEXT"],
-      policyProvided: ["YES"],
-      confirmPricipleIntegrityAndDiscretion: ["true"],
-      doneDpia: ["NO"],
-      dataDownload: ["NO"],
-      purposePursuit: ["MERE_CORRECTNESS"],
-      checkedExistenceMereCorrectnessInteropCatalogue: ["true"],
-      usesThirdPartyData: ["NO"],
-      declarationConfirmGDPR: ["true"],
-    },
-    uncompleted: {
-      purpose: ["INSTITUTIONAL"],
-      institutionalPurpose: ["test"],
-      legalBasis: ["CONSENT"],
-      knowsDataQuantity: ["NO"],
-      deliveryMethod: ["CLEARTEXT"],
-      policyProvided: ["YES"],
-      confirmPricipleIntegrityAndDiscretion: ["true"],
-      doneDpia: ["NO"],
-      dataDownload: ["NO"],
-      purposePursuit: ["MERE_CORRECTNESS"],
-      checkedExistenceMereCorrectnessInteropCatalogue: ["true"],
-      usesThirdPartyData: ["NO"],
-      declarationConfirmGDPR: ["true"],
-    },
-  },
-};
-
-export async function getRiskAnalysis({
-  tenantType,
-  completed,
-}: {
-  tenantType: TenantType;
-  completed?: boolean;
-}) {
-  const templateType =
-    tenantType === "PA1" || tenantType === "PA2" ? "PA" : "Privato/GSP";
-  const templateStatus = completed ?? true ? "completed" : "uncompleted";
-
-  const answers = RISK_ANALYSIS_DATA[templateType][templateStatus];
-
-  const token = await getToken(tenantType);
-  const { version } =
-    await dataPreparationService.retrieveRiskAnalysisConfiguration(token);
-
-  return {
-    name: "finalità test",
-    riskAnalysisForm: {
-      version,
-      answers,
-    },
-  };
-}
 
 export const getRandomInt = () =>
   Number(Math.random() * Number.MAX_SAFE_INTEGER).toFixed(0);
